requirements.txt
dash
pandas
dash_bootstrap_components
plotly.express
